test: added WIP training functionality
it appears to be working but it's late and i can't test it fully.
surprisingly faster than my previous RNN implementation but not sure if
that is because the batch size is very small, or it may be something
else.

need to implement checkpointing and saving next, and maybe play around
with the hyperparameters a bit more.